Cleanup the location module
Move the content of setlocation.go into location.go and move the global
variable `locations' to opBot.
